Fixed LayerNormalization eps bug (#589)

+++ include/tim/vx/ops/layernormalization.h
@@ -38,7 +38,7 @@ class LayerNormalization : public BuiltinOp {
 
  protected:
   int32_t axis_;
-  int32_t eps_;
+  float eps_;
 };
 
 }  // namespace ops
